* new alphaValue member for CView which affects the opaqueness when drawing 	* new optional virtual methods in CControlListener: controlTagWillChange and controlTagDidChange 	* added CFrame::close() method. Should be called instead of CFrame::forget() when you want make sure to detach the platform view from its parent 	* CScrollView draws its border with the background color, not with black anymore 	* focus handling change: a view which says wantsFocus() gets focus on mouse down now. A control does not set itself as focus view if beginEdit is called.

+++ trunk/vstgui/uidescription/cviewswitchcontainer.cpp
@@ -114,7 +114,7 @@ UIDescriptionViewSwitchController::UIDescriptionViewSwitchController (CViewSwitc
 //-----------------------------------------------------------------------------
 CView* UIDescriptionViewSwitchController::createViewForIndex (long index)
 {
-	if (index < templateNames.size ())
+	if (index < (long)templateNames.size ())
 	{
 		return uiDescription->createView (templateNames[index].c_str (), uiController);
 	}
@@ -168,7 +168,7 @@ void UIDescriptionViewSwitchController::valueChanged (CControl* pControl)
 	float min = pControl->getMin ();
 	float max = pControl->getMax ();
 	float norm = (value - min) / (max - min);
-	long index = std::min<long> (norm * templateNames.size (), templateNames.size ()-1);
+	long index = std::min<long> ((long)(norm * (float)templateNames.size ()), templateNames.size ()-1);
 	viewSwitch->setCurrentViewIndex (index);
 }
